backend/requirements.txt
flask
flask-cors
pillow
pdfminer.six